+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+- Add trace origin to span data ([#849](https://github.com/getsentry/sentry-go/pull/849))
+- Add ability to skip frames in stacktrace ([#852](https://github.com/getsentry/sentry-go/pull/852))
+
 ## 0.28.1
 
 The Sentry SDK team is happy to announce the immediate availability of Sentry Go SDK v0.28.1.